+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 🐛 Bug fixes
 
+- Added trailing zeroes to values for objects of type `float`
+
 ### 📚 Documentation and demo website changes
 
 ### 🔧 Internal changes